Align browser proof with polished workflows
