+++ apps/web/app/projects/[projectId]/settings/project/project-settings-page.tsx
@@ -1,0 +1,259 @@
+"use client";
+
+import { useParams } from "next/navigation";
+import { useEffect, useRef, useState } from "react";
+import { getProject, updateProject } from "@egma/platform-api/client";
+
+import { IDENTITY_CONFLICT, type Refusal } from "../../../../../lib/api.ts";
+import { roleOf } from "../../../../../lib/me.ts";
+import { platformAnswer, platformClient } from "../../../../../lib/platform-client.ts";
+import type { ProjectSettings } from "../../../../../lib/settings.ts";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Textarea } from "@/components/ui/textarea";
+
+import {
+  Field,
+  Form,
+  FormActions,
+  Help,
+  Problem,
+  Refused,
+} from "../../../../../ui/form.tsx";
+import { Failure, Loading, NotFound } from "../../../../../ui/page-state.tsx";
+import {
+  useOrganizationRead,
+  useUnsavedChanges,
+} from "../../../../../ui/settings-read.ts";
+import { useShellSession } from "../../../../../ui/shell.tsx";
+
+/**
+ * Edit project metadata with expectedRevision so concurrent saves cannot
+ * silently overwrite each other. Keep refused drafts for retry. Server-provided
+ * permission disables editing for users who cannot manage projects.
+ */
+export default function ProjectSettingsPage() {
+  const { projectId } = useParams<{ projectId: string }>();
+  return <ProjectSettingsBody projectId={projectId} />;
+}
+
+function ProjectSettingsBody({ projectId }: { readonly projectId: string }) {
+  const { me, refresh: refreshSession } = useShellSession();
+  // Null until the session read answers. An unsettled session is neither an
+  // admin nor a viewer, and claiming either would be a guess shown as a fact.
+  const role = me === null ? null : roleOf(me);
+
+  const { answer, reload } = useOrganizationRead<ProjectSettings>(
+    () => platformAnswer(getProject({ projectId }, { client: platformClient })),
+  );
+  const settled = answer?.status === "ready" ? answer.value : null;
+
+  /**
+   * Use the server's mayManageProjects result instead of duplicating role policy.
+   * Do not state a permission-refusal reason before the read resolves.
+   */
+  const mayAdminister = settled?.mayManageProjects ?? false;
+
+  const [name, setName] = useState("");
+  const [description, setDescription] = useState("");
+  const [saving, setSaving] = useState(false);
+  const [saved, setSaved] = useState(false);
+  const [refused, setRefused] = useState<Refusal | null>(null);
+
+  /**
+   * Which local draft a successful write captured.
+   *
+   * The fields stay editable while the request is active. If somebody types
+   * again after pressing Save, the confirming read must update `settled`
+   * without copying its older values over that newer draft. A number records
+   * the edit rather than comparing text, because typing and then returning to
+   * the same text still clears Saved and is still a new local action.
+   */
+  const editVersion = useRef(0);
+  const confirmingSave = useRef<{
+    readonly projectId: string;
+    readonly editVersion: number;
+  } | null>(null);
+  const [confirmingRead, setConfirmingRead] = useState(false);
+
+  useEffect(() => {
+    if (settled === null) return;
+
+    const confirming = confirmingSave.current;
+    confirmingSave.current = null;
+    setConfirmingRead(false);
+    if (
+      confirming?.projectId === projectId &&
+      editVersion.current !== confirming.editVersion
+    ) {
+      return;
+    }
+
+    setName(settled.name);
+    setDescription(settled.description ?? "");
+  }, [projectId, settled]);
+
+  useEffect(() => {
+    if (answer?.status === "signed-out") window.location.replace("/sign-in");
+  }, [answer]);
+
+  const named = name.trim() !== "";
+  /** Said only to somebody the server would refuse, and only once it has said so. */
+  const whyNot =
+    mayAdminister || role === null
+      ? undefined
+      : `Your ${role} role cannot change project settings. Ask an organization admin.`;
+  const changed =
+    settled !== null &&
+    (name.trim() !== settled.name ||
+      description.trim() !== (settled.description ?? ""));
+  const confirming = confirmingSave.current;
+  const changedWhileConfirming =
+    confirmingRead &&
+    confirming?.projectId === projectId &&
+    editVersion.current !== confirming.editVersion;
+  useUnsavedChanges((changed || changedWhileConfirming) && !saving, saving);
+
+  async function save(): Promise<void> {
+    if (!mayAdminister || settled === null || !named || !changed || saving) return;
+    const submittedEditVersion = editVersion.current;
+    setRefused(null);
+    setSaved(false);
+    setSaving(true);
+
+    const written = await platformAnswer(
+      updateProject(
+        {
+          projectId,
+          name: name.trim(),
+          // The slug remains part of the server contract and existing links.
+          // It is not an authoring choice on this screen, so every save sends
+          // back the exact value that was read instead of inventing or clearing
+          // one from hidden form state.
+          slug: settled.slug,
+          description: description.trim(),
+          expectedRevision: settled.revision,
+        },
+        { client: platformClient },
+      ),
+    );
+
+    setSaving(false);
+    if (written.status === "signed-out") {
+      window.location.replace("/sign-in");
+      return;
+    }
+    if (written.status !== "ready") {
+      // The typing stays exactly where it is. A refusal that cleared the form
+      // would make somebody retype their work to find out whether the second
+      // attempt fails the same way.
+      setRefused(written.refusal);
+      return;
+    }
+    confirmingSave.current = {
+      projectId,
+      editVersion: submittedEditVersion,
+    };
+    setConfirmingRead(true);
+    setSaved(editVersion.current === submittedEditVersion);
+    refreshSession();
+    reload();
+  }
+
+  if (answer === null) {
+    return <Loading what="this project" />;
+  }
+
+  if (answer.status !== "ready") {
+    return answer.status === "missing" ? (
+              <NotFound message={answer.refusal.message} />
+    ) : (
+              <Failure
+                message={
+                  answer.status === "signed-out"
+                    ? "Your session has ended. Sign in and try again."
+                    : answer.refusal.message
+                }
+                onRetry={reload}
+              />
+    );
+  }
+
+  return (
+    <>
+          {/* One form needs no extra section heading beneath the page title. */}
+          <div className="flex flex-col gap-4">
+            {refused === null ? null : (
+              <Refused
+                message={refused.message}
+                action={
+                  refused.error === IDENTITY_CONFLICT ? (
+                    <Button
+                      type="button"
+                      variant="secondary"
+                      onClick={reload}
+                    >
+                      Read this project again
+                    </Button>
+                  ) : undefined
+                }
+              />
+            )}
+
+            <Form onSubmit={() => void save()}>
+              <Field label="Name" htmlFor="project-name">
+                <Input
+                  id="project-name"
+                  value={name}
+                  autoComplete="off"
+                  spellCheck={false}
+                  disabled={!mayAdminister}
+                  aria-invalid={name.trim() === "" ? true : undefined}
+                  onChange={(event) => {
+                    editVersion.current += 1;
+                    setName(event.target.value);
+                    setSaved(false);
+                  }}
+                />
+              </Field>
+
+              <Field
+                label="Description"
+                htmlFor="project-description"
+                hint="Optional. What this project is for, for whoever opens the selector next."
+              >
+                <Textarea
+                  id="project-description"
+                  value={description}
+                  rows={3}
+                  disabled={!mayAdminister}
+                  onChange={(event) => {
+                    editVersion.current += 1;
+                    setDescription(event.target.value);
+                    setSaved(false);
+                  }}
+                />
+              </Field>
+
+              {named ? null : (
+                <Problem>A project needs a name.</Problem>
+              )}
+              {saved && refused === null ? (
+                <Help>Saved. Everybody in this organization sees the new name.</Help>
+              ) : null}
+
+              <FormActions>
+                <Button
+                  type="submit"
+                  disabled={!mayAdminister || !named || !changed}
+                  busy={saving}
+                  {...(whyNot === undefined ? {} : { why: whyNot })}
+                >
+                  {saving ? "Saving…" : "Save project"}
+                </Button>
+              </FormActions>
+            </Form>
+          </div>
+    </>
+  );
+}
